bugfix and create database again

+++ School Manager.Core/Services/Implemetations/UserService.cs
@@ -76,7 +76,7 @@ namespace School_Manager.Core.Services.Implemetations
 		}
 
 
-		public int SaveUser(UserCreateDTO User)
+		public long SaveUser(UserCreateDTO User)
 		{
 			try
 			{
